Move eval/query utils to a new module
